Fix codegate version for continue with DeepSeek
`continue version` didn't work with DeepSeek because for some reason
they enclose the user message in a sort of an envelope.

To make the change a little future-proof, we use the envelope to detect
Continue as client and then in the cli command try to extract from the
envelope, but fall back to the generic plain regex.